# docs/swagger.yaml
basePath: /
definitions:
...
  domain.Post:
    properties:
      attachment:
        $ref: '#/definitions/domain.PostAttachment'
      authorId:
        type: string
      content:
        type: string
      createdAt:
        type: string
      id:
        type: string
      pseudonym:
        type: string
      replyCount:
        type: integer
      score:
        type: integer
    type: object
  domain.PostAttachment:
    properties:
      contentType:
        type: string
      createdAt:
        type: string
      fileName:
        type: string
      id:
        type: string
      postId:
        type: string
      size:
        type: integer
    type: object
...
  handler.createPostRequest:
    properties:
      content:
        maxLength: 280
        type: string
    required:
    - content
    type: object
...
  handler.deletePostRequest:
    properties:
      id:
        type: string
    required:
    - id
    type: object
...
  handler.errorResponse:
    properties:
      code:
        type: string
      error:
        type: string
      fields:
        additionalProperties:
          type: string
        type: object
    type: object
  handler.getPostRequest:
    properties:
      id:
        type: string
    required:
    - id
    type: object
...
  handler.okResponse:
    properties:
      ok:
        type: boolean
    type: object
...
  handler.searchPostsRequest:
    properties:
      limit:
        type: integer
      query:
        type: string
    required:
    - query
    type: object
  handler.searchPostsResponse:
    properties:
      posts:
        items:
          $ref: '#/definitions/domain.Post'
        type: array
    type: object
...
info:
  contact: {}
  description: Anonymous microblogging API for posts, replies, reactions, feeds, and
    moderation.
  title: Echo API
  version: "1.0"
paths:
...
  /posts:
    post:
      consumes:
      - application/json
      parameters:
      - description: Post payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.createPostRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Post'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Create post
      tags:
      - posts
  /posts/delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: Post ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.deletePostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.okResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Delete post
      tags:
      - posts
  /posts/get:
    post:
      consumes:
      - application/json
      parameters:
      - description: Post ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.getPostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Post'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Get post by ID
      tags:
      - posts
...
  /posts/search:
    post:
      consumes:
      - application/json
      parameters:
      - description: Search payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.searchPostsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.searchPostsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Search posts
      tags:
      - posts
...
securityDefinitions:
  BearerAuth:
    description: Bearer <token>
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"